+++ src/lipsum.hpp
@@ -18,7 +18,6 @@
  * @example Markdown.cpp
  * @example Misc.cpp
  * @example ParserStressTest.cpp
- * @example Source.cpp
  */
 
 #pragma once